ingestion/requirements.txt
confluent-kafka>=2.3.0
websocket-client>=1.6.0
